+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog of IARE
 
+## v0.2.24.16
+- Book References Chart
+
+## v0.2.24.15
+- uses IARI v 4.2.2
+- WAYBACK link status now cached
+
 ## v0.2.24.14
 - Chart captions and section headings consistent markup